ui: Use data-source component for service show/instance/routing views
This also continues to migrate our steps over to steps that don't rely
on ember routes blocking and resolving before a test moves on i.e. uses
our pauseUntil spinner/wait function.

We also slightly improve the API of our pauseUntil function here. We
figured that there are three thing you might need to do within this
function.

1. retry (wait another few milliseconds before 'looking again')
2. resolve (a successful test)
2. reject (an unsuccesful test)

so we added a third argument here called 'retry' which just wraps
Promise.resolve which is what we are currenly using. Adding a specific
'retry' function makes this easier to use. We aren't using this yet, but
we'll probably move all our pauseUNtil calls to use this retry function
insted of Promise.resolve.